Prepare infrastructure to gradually absorb the upstream sourcetree
