+++ .gitignore
@@ -17,6 +17,7 @@
 .tmp
 *.iml
 /ddprof-stresstest/jmh-result.*
+/ddprof-lib/src/main/cpp-external/**/*
 
 **/.resources/
 
@@ -32,4 +33,3 @@ datadog/maven/resources
 
 # cursor AI history
 .history
-
