+++ ddprof-lib/build.gradle
@@ -13,6 +13,7 @@ def libraryName = "ddprof"
 description = "Datadog Java Profiler Library"
 
 def component_version = project.hasProperty("ddprof_version") ? project.ddprof_version : project.version
+def branch_lock = "jb/truncated_stacktraces"
 
 
 // this feels weird but it is the only way invoking `./gradlew :ddprof-lib:*` tasks will work
@@ -90,6 +91,41 @@ tasks.register('copyExternalLibs', Copy) {
   }
 }
 
+def cloneAPTask = tasks.register('cloneAsyncProfiler') {
+  description = 'Clones async-profiler repo if directory is missing'
+  doFirst {
+    def targetDir = file("${projectDir}/build/async-profiler")
+    if (!targetDir.exists()) {
+      println "Cloning missing async-profiler git subdirectory..."
+      exec {
+        commandLine 'git', 'clone', '--branch', branch_lock, '--depth', '1', 'https://github.com/datadog/async-profiler.git', targetDir.absolutePath
+      }
+    } else {
+      println "async-profiler git subdirectory already present."
+    }
+  }
+}
+
+def initSubrepoTask = tasks.register('initSubrepo', Copy) {
+  description = 'Copy shared files'
+  from("${projectDir}/build/async-profiler/src") {
+    include "arch.h"
+    include "incbin.h"
+    include "mutex.h"
+    include "mutex.cpp"
+    include "stackWalker.h"
+    include "stackWalker.cpp"
+    include "stackFrame*.cpp"
+    include "trap.h"
+    include "trap.cpp"
+  }
+  into "${projectDir}/src/main/cpp-external"
+}
+
+initSubrepoTask.configure {
+  dependsOn cloneAPTask
+}
+
 tasks.register('assembleAll') {}
 
 // use the build config names to create configurations, copy lib and asemble jar tasks
@@ -165,6 +201,7 @@ tasks.whenTaskAdded { task ->
             toolChain = task.toolChain
             targetPlatform = task.targetPlatform
             includes task.includes
+            includes project(':ddprof-lib').file('src/main/cpp-external').toString()
             includes project(':ddprof-lib').file('src/main/cpp').toString()
             includes "${javaHome()}/include"
             includes project(':malloc-shim').file('src/main/public').toString()
@@ -264,7 +301,9 @@ tasks.withType(LinkSharedLibrary).configureEach {
 library {
   baseName = "javaProfiler"
   source.from file('src/main/cpp')
+  source.from file('src/main/cpp-external')
   privateHeaders.from file('src/main/cpp')
+  privateHeaders.from file('src/main/cpp-external')
 
   // aarch64 support is still incubating
   // for the time being an aarch64 linux machine will match 'machines.linux.x86_64'
@@ -339,6 +378,10 @@ gradle.projectsEvaluated {
     def copyReleaseLibs = tasks.findByName("copyReleaseLibs")
     if (javadocTask != null && copyReleaseLibs != null) {
       javadocTask.dependsOn copyReleaseLibs
+    }
+    def initTask = tasks.findByName("initSubrepo")
+    if (initTask != null) {
+      compileTask.dependsOn initTask
     }
   }
 }
@@ -435,4 +478,8 @@ tasks.withType(AbstractPublishToMaven).configureEach {
   rootProject.subprojects {
     mustRunAfter tasks.matching { it instanceof VerificationTask }
   }
+}
+
+clean {
+  delete "${projectDir}/src/main/cpp-external"
 }
