+++ ddprof-lib/settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = "JavaProfiler"
+
 include ':ddprof-lib:benchmarks'
