+++ ddprof-lib/src/main/cpp/context.h
@@ -17,7 +17,7 @@
 #ifndef _CONTEXT_H
 #define _CONTEXT_H
 
-#include "arch.h"
+#include "arch_dd.h"
 #include "arguments.h"
 #include "os.h"
