+++ ddprof-lib/src/main/cpp/flightRecorder.h
@@ -22,7 +22,7 @@
 #include <limits.h>
 #include <string.h>
 
-#include "arch.h"
+#include "arch_dd.h"
 #include "arguments.h"
 #include "buffers.h"
 #include "counters.h"
