+++ ddprof-lib/src/main/cpp/os.h
@@ -17,7 +17,7 @@
 #ifndef _OS_H
 #define _OS_H
 
-#include "arch.h"
+#include "arch_dd.h"
 #include <signal.h>
 #include <stddef.h>
 #include <sys/types.h>
